MAJ de AI.cpp - correction des warning

+++ src/shared/ai/AI.cpp
@@ -217,25 +217,25 @@ namespace ai{
             adjacent_Cells.push_back(init_j);
         }
         // l'adjacente n°2 ?
-        if (moteur.getState().getCharacters()->verifValiditeCase(init_i - 1,init_j + 1) && init_i > 0 && init_j < moteur.getState().getCharacters()->getWidth() &&  moteur.getState().getCharacters()->isOccupiedByOpp(init_i - 1, init_j + 1, moteur.getPlayer(1).get()))
+        if (moteur.getState().getCharacters()->verifValiditeCase(init_i - 1,init_j + 1) && init_i > 0 && init_j < (int)(moteur.getState().getCharacters()->getWidth()) &&  moteur.getState().getCharacters()->isOccupiedByOpp(init_i - 1, init_j + 1, moteur.getPlayer(1).get()))
         {
             adjacent_Cells.push_back(init_i - 1);
             adjacent_Cells.push_back(init_j + 1);
         }
         // l'adjacente n°3 ?
-        if (moteur.getState().getCharacters()->verifValiditeCase(init_i,init_j+1) && init_j < moteur.getState().getCharacters()->getWidth() && moteur.getState().getCharacters()->isOccupiedByOpp(init_i, init_j + 1, moteur.getPlayer(1).get()))
+        if (moteur.getState().getCharacters()->verifValiditeCase(init_i,init_j+1) && init_j < (int)(moteur.getState().getCharacters()->getWidth()) && moteur.getState().getCharacters()->isOccupiedByOpp(init_i, init_j + 1, moteur.getPlayer(1).get()))
         {
             adjacent_Cells.push_back(init_i);
             adjacent_Cells.push_back(init_j + 1);
         }
         // l'adjacente n°4 ?
-        if (moteur.getState().getCharacters()->verifValiditeCase(init_i + 1,init_j) && init_i < moteur.getState().getCharacters()->getHeight() && moteur.getState().getCharacters()->isOccupiedByOpp(init_i + 1, init_j, moteur.getPlayer(1).get()))
+        if (moteur.getState().getCharacters()->verifValiditeCase(init_i + 1,init_j) && init_i < (int)(moteur.getState().getCharacters()->getHeight()) && moteur.getState().getCharacters()->isOccupiedByOpp(init_i + 1, init_j, moteur.getPlayer(1).get()))
         {
             adjacent_Cells.push_back(init_i + 1);
             adjacent_Cells.push_back(init_j);
         }
         // l'adjacente n°5 ?
-        if (moteur.getState().getCharacters()->verifValiditeCase(init_i + 1,init_j - 1) && init_j > 0 && init_i < moteur.getState().getCharacters()->getHeight() && moteur.getState().getCharacters()->isOccupiedByOpp(init_i + 1, init_j - 1, moteur.getPlayer(1).get()))
+        if (moteur.getState().getCharacters()->verifValiditeCase(init_i + 1,init_j - 1) && init_j > 0 && init_i < (int)(moteur.getState().getCharacters()->getHeight()) && moteur.getState().getCharacters()->isOccupiedByOpp(init_i + 1, init_j - 1, moteur.getPlayer(1).get()))
         {
             adjacent_Cells.push_back(init_i + 1);
             adjacent_Cells.push_back(init_j - 1);
